Simple local avatar support
If the property web.allowGravater is true and web.localAvatarUrl is not empty,
then avatar images are not requested from gravatar.com, but from another
service defined by the web.localAvatarUrl property.  The image is
requested via HTTP GET from the address defined by:

String.format("%s%s-%d.jpg", web.localAvatarUrl, email_username, width)

where email_username is everything to the left of the '@' sign, if
present, in the supplied username argument.  If no '@' sign is present,
the entire username is used.  It appears that the Gitblit code always
requests with the user's e-mail address.